backend/app/requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.2
numpy==1.24.4